refactor(core): typed ConsensusApplyResult, split peer_scoring, util extraction
