+++ src/core/conversation_plugins.rs
